+++ CHANGELOG.md
@@ -11,6 +11,7 @@ changes are currently planned.
 This version is pinned strictly to `b2-sdk-python==1.17.0` for the same reason.
 
 ### Added
+* Add `license` command, for printing licenses of b2 and libraries
 * Add `--write-buffer-size` parameter
 * Add `--skip-hash-verification` parameter
 * Add `replication-status` command